rework, compute lden on finaliseReceiver step
